Dice database commands
Dice expressions can now be stored in the database in the same fashion as
Characters. (NB: database entities are 'Roll' to avoid the potentially
awkward singular version of 'dice' -> 'die' in some commands).
Roll names are not unique, but can be differentiated by the owner's name
and discriminator.
The roll command now also accepts Roll names, and will roll the saved
expression if one is present and has been created by the current user.

Changes:
 - Create 'rolls' database table
 - Create the following commands:
   - add-roll
   - delete-roll
   - view-roll
   - clear-rolls
 - Update 'roll' command to accept a roll alias
 - Create transformers util
 - Update 'view-character' command to use transformers
   util
 - Update README